refactor: split Agentforce poller into focused modules
Break the 1400-line script into config, net, Salesforce API, OTel mapping,
policy, and ledger modules. Pin and discovery share process_session.
